Allow two agents to run on the same cluster
The change opens up the ability to run two or more agents pointed to
different fleet-controllers. The specific use case where this is useful
is for fleet managing a cluster that is running fleet to manage other
clusters. More concretely we have this with Rancher managing Rancher.
In this situation you want two fleet agents, one that is running for
fleet-local and one that is run for the upstream fleet-controller.