Add ability to update gitjob with branch in webhook payload
Today in gitjob's webhook design, it takes incoming webhook payload and
find the gitjob that matches the repo url regarding of which branch the
commit is coming from. This will cause a problem where a single webhook
event will trigger update on every branch of gitjob. This commit adds
the ability to read branch/tag from payload and only update gitjob that
matches branch in the payload. It also adds the ability to read from tag
so that tag event is separated from push event.