Add test for gitrepo updates
The test force pushes into the configured branch and expects an update
to the deployment.

This test uses a "dummy" repo in a separate github account.
The workflow requires a valid deploy key for that repo.
RSA keys did not work while testing.

The following Github repository secrets need to be set to (ecdsa) deploy
keys:

* CI_SSH_KEY
* CI_SSH_PUBKEY