Skip pruneBundlesNotFoundInRepo in output mode
Before https://github.com/rancher/fleet/pull/764 "fleet apply -o" would
result in a dry run, that did not reach out to a cluster.
The prune PR added pruned objects to the output, but needs to list the
bundles to do so.

The tests in fleet-examples use the output mode and run without any
cluster.

After this PR, if fleet apply is in output (= dry run) mode, it won't
contact a cluster. However it can no longer output which bundles it
would prune.