Removing ignored paths for CI runs
The CI test is a required job for merging pull requests. If there
is a change that does not trigger this CI job the PR can't be
merged as it does not meet branch protection rules. There are
cases where this can happen.

While it is extra computing work to run these jobs when there are
no changes to the code there is additional process overhead to
deal with PRs that don't meet branch protection rules and who
has access to merge those.

Signed-off-by: Matt Farina <matt.farina@suse.com>