Setting the Kubernetes libs at 0.24.5
0.24.5 is the latest on patch version for the 1.24 client libraries.
Rancher supported Kubernetes 1.24 as the latest version and those
libraries are using inside of Rancher itself. Setting to 1.24
libraries so that the chain of dependencies does not force Rancher
to use the 1.25 libraries.

Note, before this fleet was using 0.24.3, 0.25.0, and 1.24.4. It
was inconsistent. Dependabot to update Kubernetes packages can
lead to inconsistent versions. It's often useful to update them
all together as part a manual change.

Signed-off-by: Matt Farina <matt.farina@suse.com>
(cherry picked from commit 97dc124e178839a7864afc651780b8d908200b09)