Exit early for K8s 1.24 in ClusterRegistrationToken handler
Kubernetes 1.24 doesn't populate serviceAccount.Secrets field anymore,
so `len(sa.Secrets)` is always 0.
However if we get the secret and it has a token, it's safe to continue.
We don't need to reload or Enqueue() to wait for the token.
It's unfortunate that on new clusters, the `len(sa.Secrets) == 0` case is
in the happy path, but it's compatible with old clusters.