Stop requeuing on failed deployments
This change fixes an issue where failed bundle deployments
constantly requeued. This caused some problems including:

1. The constant requeuing put considerable load on the k8s API as
   it constantly looks up the Helm history
2. In some situations each retry would increment the Helm release

This change captures certain errors and turns them into status
updates instead of returning errors. This avoids constant
requeuing.

A new condition was added to achieve this. Deployed and Monitored
are automated by generated code. When DeployBundle() returns an
error Deployed is set to False. Otherwise it's true. MonitorBundle()
works the same for Monitored. An Installed Condition was added that
is manually managed. It contains the error that is propogated from
BundleDeployment to Bundle and Gitrepo. Status is also updated to
work with the display of the message.

Note, every time there is an install or upgrade it is "installed"
twice. Each time the Helm release is incremented. There is a future
optimization that can happen around this.

Relates to #499 (Fixes it)

Signed-off-by: Matt Farina <matt.farina@suse.com>