Reading contents&secrets uses specific service account
The RBAC previously allowed any user to get contents and any secrets in
the system registration namespace.
However, their names are random and listing was not allowed.

The agents on downstream clusters already use the 'request-' account to
retrieve bundledeployments from the upstream cluster and report back the
status. We extend that account to be able to retrieve 'contents', too.

The agents use a 'import-' account for reqstration, e.g. to create a
ClusterRegistration. That one is extended, so it can read the secrets in
the agent's cluster registration namespace.