This commit fixes an issue preventing versions and repos in Helm specs from being overridden in target customizations.
To do this, the repo, chart and version properties of the top level Helm Spec of a bundle are propagated to target customizations.
This is required to allow Fleet to retrieve the overridden Helm charts and include them in the bundle.
This commit also propagates these properties in bundle deployments.
For additional context, see: https://github.com/rancher/fleet/pull/1011