Returning nil on signal exit
When a SIGTERM was sent it caused the context to complete. The context
completing due to canceling is recorded as an error. Returning the
context Error returned an error which triggered a non-zero exit code.
This is not the intent. Returning nil instead. Nil is what is returned
by other wrangler users for this situation and reason.

This context is only canceled on a signal. So, it should be safe to
ignore the error on this context.

Relates to #1096

Signed-off-by: Matt Farina <matt.farina@suse.com>