Fix Rancher upgrade
CattleNamespaceMigrated Cluster status field is true in Rancher v2.5.16 even though it is still using the old namespace fleet-system. That's a problem when migrating to the new namespace cattle-fleet-system as downstream clusters thinks migration already happened, and registration was already done. That is not true, and it's missing the fleet-agent secret that contains the kubeconfig in the downstream cluster. Therefore, new downstream fleet-agents can't connect to the upstream cluster.

Old agent is not removed in downstream clusters because cluster.Spec.AgentNamespace is always empty. cluster.Spec.AgentNamespace just changes for the local fleet agent as it is passed as a bootstrap config when installing it in Rancher.

Set CattleNamespaceMigrated to false if the agent status namespace is fleet-system, so the new fleet-agent is registered. Remove cluster.Spec.AgentNamespace should be empty condition to delete old agents