Long release names were shortened twice
Long release names are shortened and a md5 hash is appended. The
existing code would shorten a release name twice. Once when creating the
bundle name and again when creating the helm release name.
This happened because name.Limit(str, 53) shortens any string smaller
than 53 chars to 53 chars. It will modify a string of 53 chars, so it's
not stable when calling it multiple times.

When the release name is different from the bundle name, fleet will
clean up such a release as it no longer belongs to a bundle.