Watch for a specific namespace in the BundleDeployment controller in the integration tests
Add a NewSharedControllerFactory with a default namespace, so the
BundleDeployment controller will just watch resources in the given
namespace. This is what the fleet agent is doing, and it allows multiple
tests to run in different namespaces.

Setup integration specs in suite, doesn't work if moved into the spec's
`BeforeAll()`.

Co-Authored-By: raul <raul.cabello@suse.com>