Split examples tests from e2e tests
The examples need to be tested, because the fleet-examples repo is
part of our documentation. This PR moves the tests into an acceptance
folder.
The tests are highly repetetive and we don't expect them to provide much
coverage. However, we want to test more "end-user" applications in the
future, e.g. when a new version of cert-manager comes out, we want to
know if fleet can still install it.

Acceptance test are only run on schedule, because they have limited
value for validating PRs.

For validating PRs we keep using the e2e tests. They focus more on
features, like OCI helm charts, and should provide high coverage of
the fleet code.