Retry helm upgrade if it was interrupted
When a helm upgrade is interrupted it returns the error: another operation (install/upgrade/rollback) is in progress. Then the release is in a pending-upgrade state. This happens if the fleet-agent pod is killed while doing an upgrade.

To solve this we do a rollback if this error was found, then retry the upgrade.

Signed-off-by: raul <raul.cabello@suse.com>