Fix helm repository creation in integration tests
"address already in use" error was returned in very rare occasions. This was due to the previous helm repository was not fully closed before creating the new one, as it might take an extra milliseconds to fully close after calling closed.

Retry if "address already in use" error is returned after 100 milliseconds for a maximum of 10 attempts

Signed-off-by: raul <raul.cabello@suse.com>