Set up local Helm registry
As a preliminary step to making OCI tests more portable, this creates a
local Helm registry via Zot [1], to which charts can be pushed from
outside the cluster.

Since TLS is needed to enable pushing, this commit includes a script to
generate certificates, which are then populated in Zot configuration
whereas the corresponding root certificate needs to be installed on the
host.

[1]: https://zotregistry.io/v1.4.3/