Allow helm credentials to be defined for each path
Add HelmSecretNameForPaths in GitRepo

This allows to specify different credentials for each Bundle that will be generated for each path.
HelmSecretName will be ignored if HelmSecretNameForPaths is present.

Tests returned an "address already in use" error on very rare occasions. This was due to the previous helm repository was not fully closed before creating the new one, as it might take an extra milliseconds to fully close after calling closed. Retry if "address already in use" error is returned after 100 milliseconds for a maximum of 10 attempts