Add webhook-based test case for GitRepo
This provides a new test case for GitRepo using webhooks instead of
polling.
Instead of using an external git hosting provider (eg. Github) to
configure a webhook and trust them to deliver it, we leverage a local
git server. On that server, we configure a _post-receive_ hook which,
upon each push, will call the `GitJob` service with an HTTP request body
and headers simulating a Github webhook delivery.