Prevent redundant test git server image rebuilds
The `dev/import-images-k3d` script now checks for existence of the
`nginx-git:test` Docker image used to set up a test git server.
That image will only be built if it does not already exist or if a new
`FORCE_GIT_SERVER_BUILD` environment variable is set to a non-empty
value.