Drift correction
Remove any external change when drift is detected. This is done by doing a rollback to the latest release when drift is detected. Add CorrectDrift options to GitRepo, Bundle and BundleDeployment.
Drift will be corrected just if correctDrift.enabled is true.
Helm rollback with --force is used if correctDrift.force is true.
Failed rollbacks will be removed from history unless correctDrift.keepFailHistory is true.

Signed-off-by: raul <raul.cabello@suse.com>