Add GitRepo e2e test case involving auth to a local Helm registry (#1591)
* Make Fleet test config script compatible with custom cluster name

This adds a cluster name parameter, defaulting to `upstream`, to
`dev/setup-fleet`.

* Set up local Helm registry

As a preliminary step to making OCI tests more portable, this creates a
local Helm registry via Zot [1], to which charts can be pushed from
outside the cluster.

Since TLS is needed to enable pushing, this commit includes a script to
generate certificates, which are then populated in Zot configuration.

[1]: https://zotregistry.io/v1.4.3/

* Add port number support to OCI chart downloads

The chart URL specified in `fleet.yaml` may contain a port number, in
which case that port number must be taken into account when trying to
log into the corresponding registry. Failing that, connection attempts
may fail and be visible as timeouts.

* Add e2e OCI test case using a local Helm registry

At this point, pushing to the registry still prompts for a passphrase on
my machine, and a `fleet` container complains about a `certificate signed
by unknown authority`. That issue may be related to [fleet#1508][1].

[1]: https://github.com/rancher/fleet/issues/1508

* Install CA cert only if present

An attempt to mount a CA cert into a Fleet agent pod must be made only
if that cert is provided in the Helm secret configured for the bundle.

* Support self-signed certs for path-specific Helm credentials

This extends support for root CA certificates specified in Helm secrets
to path-specific secrets.

* Ignore `-passin` openSSL flag when spell checking

That flag is not a misspelt version of `passing`.

* Add OCI env vars to e2e CI workflow

Additional environment variables are needed to enable secret creation
and interaction with the local Helm registry from the new OCI test case.

* Skip git server image building if env variable not defined

When environment variable `FORCE_GIT_SERVER_BUILD` is not set, then it
is considered to have an empty value. This prevents
`dev/import-images-k3d` from complaining about the variable not being
set in such cases.

* Decouple infra setup and teardown from e2e test suite

Infrastructure needed by the end-to-end test suite can now be set up and
torn down separately, via new commands implemented in
`e2e/testenv/infra`:
* `infra setup` spins up a git server and a Helm registry in the
  test cluster, before pushing a test Helm chart to that registry
* `infra teardown` destroys the git server, Helm registry and associated
  Kubernetes resources.

* Document e2e test environment setup and teardown steps

This updates the dev README with new steps needed to set up and tear
down and e2e test environment, now that installing a root certificate on
the host is no longer needed.

* Avoid test certs generation if test namespace not set

When environment variable `FLEET_E2E_NS` is not set, certificates would
be generated with an unusable SAN DNS name, which would ultimately lead
to OCI test failures. Therefore, the certificates generation script now
exits in such cases.

* Add test setup binary and chart to .gitignore

Those files would only clutter the repository.

* Wait for test environment to be ready

Instead of using blind `Sleep()` statements, the e2e test environment
setup script now checks git server and Helm registry pods for readiness.
Additionally, since that check is not enough in the latter case, this
adds a retry mechanism with a timeout for logging into the Helm
registry.

* Remove e2e test env setup and teardown code from imagescan tests

This is now handled via a separate script.

* Export certs dir via dedicated command

Github actions `env` key does not seem to support command substitution.

* Provide optional Helm path override

This enables both bypassing Rancher's Helm fork, which is outdated at
the time of writing, and pointing the CI worker to a Helm executable.

* Enable multiple consecutive OCI test runs

This prevents false positives when re-running OCI tests involving a
local Helm registry, by checking for _running_ `sleeper` pods only
instead of including terminating ones left over from previous runs.

Besides, runs now delete the `helm` gitrepo, for a more complete
cleanup. When only deleting the `sleeper` deployment, subsequent runs
would show gitrepo or bundle errors about a missing `sleeper` app.

* Remove external IP from Zot cert SAN IPs

Logging into and pushing a Helm chart to a local Zot instance with TLS
enabled is now done using flags `--insecure` and
`--insecure-skip-tls-verify`, respectively. This eliminates the need to
have Zot's external IP, used for those operations from outside the
cluster, featured in a certificate.