Add ChartMuseum to e2e tests infra setup and teardown commands
This enables ChartMuseum, needed to test retrieving Helm charts via HTTP
URLs, to be set up and torn down automatically as Zot and the git server
are.
Since two different tools are now needed to serve Helm charts, this
commit does some renaming to clarify that resources are related to
serving Helm charts, not just exposing an OCI registry.