Add NamespaceLabels and NamespaceAnnotations
This adds the ability to append labels or annotations to the namespace created by fleet as part of the helm install process.

NamespaceLabels and NamespaceAnnotations are pointer to maps. Nothing will be done if they are nil, which means they are not provided by the use.

Helm creates the namespace with one label whose key if name. This is the only label we preserve, other labels or annotations that are not provided in NamespaceLabels or NamespaceAnnotations will be removed.

Signed-off-by: raul <raul.cabello@suse.com>