Fix Helm release name generation to prevent double dashes
A Helm release name is generated for each bundle deployed by Fleet.
In case that name is longer than 53 characters, the maximum length of
Helm release names, it is then truncated to 46 characters and the
remaining characters are a dash followed by a hash.

This fixes an edge case where the truncated release name already ends
with a dash, leading to the final release name (in the form
`<truncated>-<hash>`) containing a double dash. Since this caused
discrepancies between deployed release names and what the Fleet agent
would expect to be deployed, Fleet would frequently delete and re-create
bundles as a result.