Prevent non-hex chars in release name hash
When a release name must be truncated to comply with Helm's maximum
release name length, Fleet truncates the name and builds a new release
name as `<name>-<hash>`.

In cases where the last character of `<name>` is a dash, a previous
commit would include the character immediately following the dash and
place it before the hash, which might cause confusion.

Instead, this commit does not include that character in such cases and
increases the hash length by 1, so that characters after the last dash
in the release name remain hexadecimal.