Avoid increasing hex length if truncated string ends with `-`
In such cases, the separator is simply omitted. The resulting string
will be one character shorter than the limit, which should not cause any
issues. On the other hand, the hash will contain only hexadecimal
digits, as it previously did.
Since double dashes are not part of the result, this preserves
consistency across repeated calls, and hence between bundle ID and Helm
release name, thereby preventing unexpected deletion and re-creation of
bundles.