Prevent non-hex chars in release name hash (#1649)
* Prevent non-hex chars in release name hash

When a release name must be truncated to comply with Helm's maximum
release name length, Fleet truncates the name and builds a new release
name as `<name>-<hash>`.

In cases where the last character of `<name>` is a dash, a previous
commit would include the character immediately following the dash and
place it before the hash, which might cause confusion.

Instead, this commit does not include that character in such cases and
increases the hash length by 1, so that characters after the last dash
in the release name remain hexadecimal.

* Avoid increasing hex length if truncated string ends with `-`

In such cases, the separator is simply omitted. The resulting string
will be one character shorter than the limit, which should not cause any
issues. On the other hand, the hash will contain only hexadecimal
digits, as it previously did.
Since double dashes are not part of the result, this preserves
consistency across repeated calls, and hence between bundle ID and Helm
release name, thereby preventing unexpected deletion and re-creation of
bundles.