Initialise cluster registration labels map if empty
This prevents attempts to add a label to a cluster registration from
panicking and making the Fleet agent crash.

This commit provides a smoke test which, while not testing the full
extent of cluster registration (which would require at least a mock
Kubernetes API server), validates that trying to register a cluster does
not panic where it used to.