Use custom release script for test Fleet release
This commit provides a new shell script which packages new Helm charts
from Fleet binaries and releases those charts to a test branch of a
charts repo.
This script is intended to be called from CI workflow
`release-against-test-charts.yml` against a fork of `rancher/charts`,
but could be called locally as well.