Support manually and PR-triggered runs
This commit ensures that the Github ref, the test charts repo and base
branch are set even when the workflow is triggered by an event, such as
a pull request.
Those values can still be overwritten when manually running the
workflow.
Besides, the previous Fleet application and chart version are not deemed
necessary at this point, as the workflow can compute them from the
latest released Fleet chart on the base charts branch.