Create non-root users and groups in Dockerfiles
This ensures that a user and a group is created in the Fleet controller
and agent Docker images for UID 1000. This prevents unexpected
behaviour, such as corresponding host `/etc/passwd` entries being
copied, as podman does.

The controller image is based on BusyBox, where `addgroup` and `adduser`
are available out of the box. Regarding the agent image, neither
`groupadd` nor `useradd` are included in
`registry.suse.com/suse/git:2.35`, nor even `zypper` which would have
enabled `pwdutils` to be installed. Therefore, we chose to switch to a
larger base image, and to remove `curl` from it to reduce false
positives from security scanners.