Switch base images to smaller images
Also moving the USER statement to the bottom. Experiments show it works
in any place, but it's confusing to read. Note, the USER statement is
the default UID if no securityContext is used. Fleet uses a
securityContext/RunAsUser, but scanners complain when USER is missing.
This is probably because scanners don't distinguish between the docker
container runtime and the container runtime used in k8s.