Add goreleaser workflow from #1019
This adds a new CI workflow for releasing Fleet via GoReleaser, with
configuration inherited from a previous pull request and extended.

This runs linters, then unit and integration tests, after which it
builds then releases all needed binaries and Docker images for Fleet.

Dockerfiles are adapted to GoReleaser's particularity of building Docker
images using the build directory as context. In this case, we configure
GoReleaser to output all binaries to a common root `dist/` directory, so
that all binaries are available to all Docker image build processes.
This is useful to include eg. both `gitjob` and `gitcloner` binaries in
the `rancher/gitjob` image.

Windows-based images are not supported, and should not be needed anyway
as the Fleet agent is only ever deployed to Linux k8s nodes.