Move fleet CLI to gitjob Docker image
The fleet CLI binary now lives in the `rancher/gitjob` image, as that
binary only needs to run on the upstream cluster.
As a result, the cluster registrations cleanup job uses that image
instead of `rancher/fleet-agent`, and so do git jobs.

The agent image now uses Busybox as its base image, which is more
lightweight.