Fix Existing Namespace Not Found By Label (#2204)
* Filter by kubernetes.io/metadata.name label

Use the label kubernetes.io/metadata.name, which is added by k8s to all
namespaces, instead of the helm "name" label. The helm "name" label
might be missing from pre-existing namespaces.

* Switch fetchNamespace to Get instead of list

Was using List with a label, that's set by k8s to the name of the
namespace. Only returned the first result.
Testdata assumed label and namespace name could be different.