Propagate bundle and bundledeployment status updates to GitRepo
A filter on status updates in the GitRepo reconciler prevents updates to
statuses of bundles and bundle deployments from being reflected in their
owning GitRepo's status. In turn, this leads to outdated and
inconsistent GitRepo statuses.

This commit remedies this, with a first attempt at covering GitRepo
status updates in end-to-end tests.