Adds option to disable polling for Git repos
Adds the `disableGitPolling` option in order to disable polling for git repos.
It also ports gitrepos polling to use the `quartz` library to be inline with
`imagescan`, that is already using it.

Changes the unit tests to be implemented using the Ginkgo testing framework.

Refers to: https://github.com/rancher/fleet/issues/1572

Signed-off-by: Xavi Garcia <xavi.garcia@suse.com>