Propagate bundle and bundledeployment status updates to GitRepo (#2223)
* Propagate bundle and bundledeployment status updates to GitRepo

A filter on status updates in the GitRepo reconciler prevents updates to
statuses of bundles and bundle deployments from being reflected in their
owning GitRepo's status. In turn, this leads to outdated and
inconsistent GitRepo statuses.

Eliminating event filters would increase the number of reconcile loops,
as each update to a `GitRepo`'s status would trigger a new loop.
Instead, this commit widens the set of events allowed via predicates to
bundle status updates.

* Compute target namespace for webhook-based gitrepo test

This was already done for polling-based gitrepo testing, but missing for
the webhook-based test case, which would lead to conflicts.