Add workflow releasing Fleet to test charts repo
This workflow builds Fleet from the specified branch, pushes Docker
images and packages charts with hard-coded version number
`999.9.9+up9.9.9`, which it edits for Rancher compatibility and pushes
to a custom branch on a rharts repository, using the specified
`rancher/charts` branch as its base.

Created Docker images are pushed to ttl.sh [1], with 1h as expiration
time.

This is meant to be eventually used by a separate workflow running tests
against Rancher. In the meantime, it can be used for manual testing,
installing Rancher with flags:
* `CATTLE_CHART_DEFAULT_URL` set to the test charts repo
* `CATTLE_CHART_DEFAULT_BRANCH` set to the branch created by this
  workflow
* `CATTLE_FLEET_VERSION` set to `999.9.9+up9.9.9`

[1]: https://ttl.sh/