Hide global CLI flags on sub commands
The binaries run a default action if no sub command is given.
The flags of the default action, e.g. running the main fleet controller,
are inherited by the sub commands as global flags.

This is a easy fix to hide the unsupported flags. We could
turn all parent commands into sub commands one day, but it seems low
priority.