Add sharding to Fleet deployments in CI workflows running end-to-end tests (#2405)
* Update nightly CI workflow for sharding and metrics

Similarly to what has been done for the regular CI workflow for
end-to-end tests, this workflow needs the following, added here:
* a Fleet deployment with sharding enabled
* a metrics tests run

* Fix Fleet deployment for AKS and GKE workflows

Now that sharding end-to-end test cases can be run without test infra,
they are run in AKS and GKE workflows. Those workflows therefore require
Fleet to be deployed with sharding enabled.