Fix multi-cluster e2e tests for strict agent TLS mode
This uses the Helm CLI to install the Fleet agent on the downstream
cluster, instead of relying on an external script which is subject to
changes.
Logic used to delete previous Fleet agent installs now deletes the
`cattle-fleet-system` namespace as well, which leaves a clean slate for
further test cases.