Honor Helm's MaxHistory when using drift correction (#2519)
Drift correction (self-healing) uses Helm's Rollback action, which has                                                                                                                                 a MaxHistory field. This value takes precedence over the config value.
Unlike the Upgrade action, we omitted to set this value, so the releases
will only shrink when running an upgrade but not rollbacks.

Regular upgrades would sometimes be deployed (creating a new release and
secret). This was caused by the BundleDeployment status having a release
field, referencing the latest release installed. This value was not                                                                                             being updated on rollbacks, hence making the status differ from what
helm history reflects.