Make drift correction end-to-end tests more robust
These tests now wait for deletion of a `GitRepo` to complete before
carrying on with the next test case, which could prevent interferences.

In case of a failure to get a bundle ready on time before running a test
case they should also now output bundle status to ease troubleshooting.