Fix flaky e2e tests (#2539)
* Wait for target namespace deletion

This may prevent a race condition whereby cleanup for a given test case
may finish deleting a namespace _after_ the next test case starts
running, leading to that test case failing to set itself up, with
`BeforeEach` not finding the required config map.

* Retry docker pulling image in imagescan tests

In some cases, imagescan end-to-end tests fail because of a failure to
pull an image. This enables retries in such cases, in the hope of
reducing or eliminating test flakiness.

* Give integration tests more time to comply with expectations

This may fix flakiness on Github runners.

* Refine error check in Fleet controller logs

Checking logs for a single `ERROR` keyword may lead to false positives,
for instance caused by other test suites. Instead, this uses a regular
expression to look for logs related to missing bundles or bundle
deployments.

* Make drift correction end-to-end tests more robust

These tests now wait for deletion of a `GitRepo` to complete before
carrying on with the next test case, which could prevent interferences.

In case of a failure to get a bundle ready on time before running a test
case, they should also now output bundle status to ease troubleshooting.

This also improves output of drift correction tests to help detect patches
which may not have caused expected drifts.

* Clean up after tests

This ensures that more namespaces and `GitRepo`s created by
end-to-end tests are deleted after execution.

* Expose error if bundle not ready

This should ease troubleshooting of flaky tests for deploying bundles
from `GitRepo`s with multiple paths.

---------

Co-authored-by: Alejandro Ruiz <alejandro.ruiz@suse.com>